flask-API-master (3)/flask-API-master/requirements.txt
alembic==1.13.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
packaging==24.0
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
tabulate==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
what3words==3.1.1
